Using a simple registry instead of a Reg registry for dump_json, load_json, and link_prefix.
