+++ morepath/directive.py
@@ -45,6 +45,10 @@ from .path import PathRegistry
 from . import generic
 from .settings import SettingRegistry
 
+try:
+    from functools import singledispatch
+except ImportError:
+    from singledispatch import singledispatch
 
 def isbaseclass(a, b):
     return issubclass(b, a)
@@ -1153,10 +1157,49 @@ class VerifyIdentityAction(dectate.Composite):
                              identity=self.identity), obj
 
 
+class SimpleRegistry(object):
+    def __init__(self):
+        self._dump_json = singledispatch(lambda obj, request: obj)
+
+    @staticmethod
+    def link_prefix(request):
+        """Returns a prefix that's added to every link generated by the request.
+
+        By default :attr:`webob.request.BaseRequest.application_url` is used.
+
+        :param request: :class:`morepath.Request`
+        :return: prefix string to add before links.
+        """
+        return request.application_url
+
+    @staticmethod
+    def load_json(request, json):
+        """Load JSON as some object.
+
+        By default JSON is loaded as itself.
+
+        :param request: :class:`morepath.Request`
+        :param json: JSON (in Python form) to convert into object.
+        :return: Any Python object, including JSON.
+        """
+        return json
+
+    def dump_json(self, request, obj):
+        """Dump an object as JSON.
+
+        ``obj`` is any Python object, try to interpret it as JSON.
+
+        :param request: :class:`morepath.Request`
+        :param obj: any Python object to convert to JSON.
+        :return: JSON representation (in Python form).
+        """
+        return self._dump_json(obj, request)
+
+
 @App.directive('dump_json')
 class DumpJsonAction(dectate.Action):
     config = {
-        'reg_registry': RegRegistry
+        'proxy': SimpleRegistry
     }
 
     filter_convert = {
@@ -1183,20 +1226,17 @@ class DumpJsonAction(dectate.Action):
         '''
         self.model = model
 
-    def identifier(self, reg_registry):
+    def identifier(self, proxy):
         return self.model
 
-    def perform(self, obj, reg_registry):
-        # reverse parameters
-        def dump(request, self):
-            return obj(self, request)
-        reg_registry.register_function(generic.dump_json, dump, obj=self.model)
+    def perform(self, obj, proxy):
+        proxy._dump_json.register(self.model)(obj)
 
 
 @App.directive('load_json')
 class LoadJsonAction(dectate.Action):
     config = {
-        'reg_registry': RegRegistry
+        'proxy': SimpleRegistry
     }
 
     def __init__(self):
@@ -1208,20 +1248,20 @@ class LoadJsonAction(dectate.Action):
         '''
         pass
 
-    def identifier(self, reg_registry):
+    def identifier(self, proxy):
         return ()
 
-    def perform(self, obj, reg_registry):
+    def perform(self, obj, proxy):
         # reverse parameters
         def load(request, json):
             return obj(json, request)
-        reg_registry.register_function(generic.load_json, load)
+        proxy.load_json = load
 
 
 @App.directive('link_prefix')
 class LinkPrefixAction(dectate.Action):
     config = {
-        'reg_registry': RegRegistry
+        'proxy': SimpleRegistry
     }
 
     def __init__(self):
@@ -1236,8 +1276,8 @@ class LinkPrefixAction(dectate.Action):
         '''
         pass
 
-    def identifier(self, reg_registry):
+    def identifier(self, proxy):
         return ()
 
-    def perform(self, obj, reg_registry):
-        reg_registry.register_function(generic.link_prefix, obj)
+    def perform(self, obj, proxy):
+        proxy.link_prefix = obj
