+++ morepath/generic.py
@@ -185,41 +185,3 @@ def permits(identity, obj, permission):
     :return: ``True`` if identity has permission for obj.
     """
     return False
-
-
-@reg.dispatch()
-def load_json(request, json):
-    """Load JSON as some object.
-
-    By default JSON is loaded as itself.
-
-    :param request: :class:`morepath.Request`
-    :param json: JSON (in Python form) to convert into object.
-    :return: Any Python object, including JSON.
-    """
-    return json
-
-
-@reg.dispatch('obj')
-def dump_json(request, obj):
-    """Dump an object as JSON.
-
-    ``obj`` is any Python object, try to interpret it as JSON.
-
-    :param request: :class:`morepath.Request`
-    :param obj: any Python object to convert to JSON.
-    :return: JSON representation (in Python form).
-    """
-    return obj
-
-
-@reg.dispatch()
-def link_prefix(request):
-    """Returns a prefix that's added to every link generated by the request.
-
-    By default :attr:`webob.request.BaseRequest.application_url` is used.
-
-    :param request: :class:`morepath.Request`
-    :return: prefix string to add before links.
-    """
-    return request.application_url
